[REQ-178] [TASK-5] Integrate async execution with /requirement command
- Updated handle_requirement() to use spawn_requirement_task() from new async executor
- Added ActiveRequirementProgress type to handle ProgressMessage enum
- Updated event loop to handle ProgressMessage variants with status indicators (○, ⠋, ●, ✗)
- Added token tracking and duration updates to progress display
- Added JoinHandle storage in App struct for task lifecycle management
- Created render_requirement_progress_new() component for displaying new progress data
- TUI remains responsive during requirement execution with live progress updates

+++ apps/tui/src/effects/dialog_animations.rs
@@ -55,6 +55,6 @@ pub fn create_dialog_close_animation(
 }
 
 fn create_timer(duration_ms: u64, interpolation: Interpolation) -> EffectTimer {
-    EffectTimer::from_ms(duration_ms, interpolation)
+    EffectTimer::from_ms(duration_ms as u32, interpolation)
 }
